n8n-to-notion/requirements.txt
streamlit==1.32.0
markdown==3.4.3
beautifulsoup4==4.12.2
Pillow==10.2.0
requests==2.31.0
